Version 3
File Operations using Win32 APIs.  Server contains source code for all the APIs that will be called by the client. Server is targeted to form a DLL which will be used by client through explicit linking of the DLL. Client uses these functions to perform file operations on a regular file such as Count number of characters, Count number of words and count number of capital and small case characters.